Split eventtrace in coreclr to match
